Update README.md
For the first challenge of the course, we implemented Semantic elements to the original HTML file. For the purpose of making it easier to follow and understand the structure of the page. By doing so we also had to update the CSS coding in order to have a fluid funcitionality of the page. Throught the challenge there was some difficulties that will help me to improve my formating and expand my knowledge with the importance of Semantic Elements.